copies of predefined rules in yaml; documentation
