+++ java/core/src/test/java/co/worklytics/psoxy/rules/YAMLRulesTestBaseCase.java
@@ -1,0 +1,21 @@
+package co.worklytics.psoxy.rules;
+
+import co.worklytics.psoxy.Rules;
+import co.worklytics.psoxy.rules.google.PrebuiltSanitizerRules;
+import lombok.SneakyThrows;
+
+/**
+ * base test case for rules defined in YAML
+ */
+public abstract class YAMLRulesTestBaseCase extends RulesBaseTestCase {
+
+
+    @SneakyThrows
+    @Override
+    public Rules getRulesUnderTest() {
+        String path = "/rules/" + getYamlSerializationFilepath() + ".yaml";
+
+        return yamlMapper.readerFor(Rules.class)
+            .readValue(PrebuiltSanitizerRules.class.getResource(path));
+    }
+}
