+++ java/core/src/test/java/co/worklytics/psoxy/rules/google/CalendarTests.java
@@ -1,6 +1,7 @@
 package co.worklytics.psoxy.rules.google;
 
 import co.worklytics.psoxy.Rules;
+import co.worklytics.psoxy.rules.JavaRulesTestBaseCase;
 import co.worklytics.psoxy.rules.RulesBaseTestCase;
 import lombok.Getter;
 import lombok.SneakyThrows;
@@ -11,7 +12,7 @@ import java.util.Arrays;
 import java.util.Collection;
 
 
-class CalendarTests extends RulesBaseTestCase {
+class CalendarTests extends JavaRulesTestBaseCase {
 
     @Getter
     final Rules rulesUnderTest = PrebuiltSanitizerRules.GCAL;
